requirements.txt
paho-mqtt==2.1.0
pyham_kiss==1.0.0
pyham_ax25==1.0.2
python-decouple==3.8
